+++ examples/hello/elements/MyTD.jsx
@@ -11,10 +11,11 @@ export default React.createClass({
   displayName: 'MyTD',
 
   propTypes: {
-    bgColor: Oy.PropTypes.rules(['SixCharacterHexBackgroundColorRule'])
+    bgColor: Oy.PropTypes.rules(['SixCharacterHexBackgroundColorRule']),
+    style: Oy.PropTypes.rules(['EmptyTDRule'])
   },
 
   render: function() {
-    return <Oy.Element type="td" {...this.props} />;
+    return <Oy.Element tagName="td" {...this.props} />;
   }
 });